+++ src/app/locales/en/components/file-browser/fb-dir-load-error.js
@@ -1,3 +1,4 @@
 export default {
   unknownError: 'Unknown error',
+  navigateToRoot: 'Navigate to root',
 };