+++ src/app/locales/en/components/filesystem-browser/dir-load-error.js
@@ -1,0 +1,3 @@
+export default {
+  navigateToRoot: 'Navigate to root directory',
+};